refactor: code separated
