+++ unity-client/Assets/Scripts/MainScripts/DCL/Controllers/BuilderMode/Catalog/FavoritesController.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 895d63441da05b6449048521e5890f74
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
