+++ unity-client/Assets/Scripts/MainScripts/DCL/Controllers/BuilderMode/Catalog/QuickBarController.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 63ca2ec1bcf6c364fbd37f25eee7e8d2
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
